bot/requirements.txt
python-telegram-bot==12.0.0b1
python-dateutil
python-telegram-bot[socks]